Manually tag containers instead of using Docker Tag github action
The Docker Tag action tags containers just fine and dandy, and
then we can upload those containers easy as pie, but that's where
it ends.

To deploy the newly-uploaded container by using a non-hardcoded tag,
we need to be able to change the tag on every merge to master:
an easy way of doing that is using the github commit hash.

The Docker Tag action tags each container with the first 7
characters of the commit hash for us, which is great,
but try as I might, I haven't managed to access that variable
in later actions.

So, I want to reference that tag to deploy it using the Azure cli
action (as opposed to "latest", which was hardcoded), but it doesn't
look like that variable is available to later actions.

So instead I am manually tagging with the GITHUB_SHA variable,
which is a github-injected environment variable which means that
it can be reused across actions.

More info on environment variables consistent across actions:
https://developer.github.com/actions/creating-github-actions/accessing-the-runtime-environment/#runtime-environment-resources